Refactor Kramdown custom extension handling
Move custom extension out to files.
